Add files via upload
SEIR_2 is a python module containing the core of data gathering and SEIR fitting. No extended documentation is yet available, this is a toy version. 
SEIR_2_test is a showcase fro SEIR_2
SEIR testing parameter estimation with known model does what the name says: the methods used in SEIR_2 are tested to recover known parameters of SEIR-generated data points